+++ docker-compose.yml
@@ -38,3 +38,16 @@ services:
     environment:
       - MYSQL_ALLOW_EMPTY_PASSWORD=yes
       - MYSQL_DATABASE=training
+  db-test:
+    build:
+      context: ./docker
+      dockerfile: mysql.Dockerfile
+    volumes:
+      - ./docker/mysql-test:/var/lib/mysql
+      - ./mysql/my.cnf:/etc/mysql/conf.d/my.cnf
+      - ./mysql/sql:/sqlscripts
+    ports:
+      - '3307:3306'
+    environment:
+      - MYSQL_ALLOW_EMPTY_PASSWORD=yes
+      - MYSQL_DATABASE=training
